app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudscraper==1.2.71
colorama==0.4.6
cssselect==1.3.0
discord.py==2.5.2
distro==1.9.0
dotenv==0.9.9
frozenlist==1.6.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
lxml==5.4.0
multidict==6.4.3
openai==0.28.0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
w3lib==2.3.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.0